Revert BedrockBase worker thread configuration (#404)
Reverts the unreleased `max_parallel_requests` attribute added in #391,
as it is redundant with the
[`max_concurrency`](https://python.langchain.com/api_reference/core/runnables/langchain_core.runnables.config.RunnableConfig.html#langchain_core.runnables.config.RunnableConfig.max_concurrency)
option already available in RunnableConfig for batching calls.

In general, the Runnable async invoke and stream methods are not
intended to be able to make multiple requests in parallel, and do not
use `max_concurrency`. We should not attempt to skirt this restriction
via a custom parameter, and instead recommend users to utilize the
purpose-built
[`abatch`](https://python.langchain.com/api_reference/core/runnables/langchain_core.runnables.base.Runnable.html#langchain_core.runnables.base.Runnable.abatch)
method.