Prevent sending thinking blocks with no signature with ChatBedrockConverse (#408)
Fixes https://github.com/langchain-ai/langchain-aws/issues/401

Claude 3.7 still works because signatures are always returned with those
thinking blocks. DeepSeek thinking blocks do not contain signatures and
the ValidationError indicates a signature must be included with thinking
blocks.

According to Converse API docs:
> "Within the reasoningText field, the text fields describes the
reasoning. The signature field is a hash of all the messages in the
conversation and is a safeguard against tampering of the reasoning used
by the model. You must include the signature and all previous messages
in subsequent Converse requests. If any of the messages are changed, the
response throws an error."


https://docs.aws.amazon.com/bedrock/latest/userguide/conversation-inference-call.html

@3coins I posted a separate PR
https://github.com/langchain-ai/langchain-aws/pull/407 to make reviewing
easier. Happy to combine if preferred.