Add base model ID field to ChatBedrock models (#431)
Fixes #228 
Fixes #253 
Fixes #394 

Added an optional `base_model_id`(or `base_model`) parameter to
BedrockBase and ChatBedrockConverse.

This provides an alternate means of identifying the base model if
`model_id` is ambiguous (i.e. a custom or provisioned model ARN). In
particular, this is useful for applying specific Bedrock request tuning
(ex. prompt templating in ChatBedrock, and streaming support levels in
ChatBedrockConverse) for different models under the same provider.