aws[patch]: support model_kwargs (#463)
Implement a convenience for initialization that is common to most other
LangChat chat models.

Here we catch params on init that are invalid according to Pydantic and
store in the `model_kwargs` parameter (or
`additional_model_request_fields` in the case of ChatBedrockConverse).

Consider initializing with a param that is not explicitly typed as an
attribute:

```python
from langchain_aws import ChatBedrock, ChatBedrockConverse

thinking = {"type": "enabled", "budget_tokens": 2000}  # invalid param

llm_1 = ChatBedrock(
    model="anthropic.claude-3-5-sonnet-20240620-v1:0",
    thinking=thinking,
)
llm_2 = ChatBedrockConverse(
    model="anthropic.claude-3-5-sonnet-20240620-v1:0",
    thinking=thinking,
)
```
Currently, both of the above raise ValidationError.

After this change, ChatBedrock will transfer `thinking` to
`model_kwargs`, and ChatBedrockConverse will transfer it to
`additional_model_request_fields` (while emitting a warning). Both will
include it in request payloads.