fix: Optimize base model identification for application inference profiles (#615)
This PR updates BedrockBase and ChatBedrockConverse to optimize the
identification of base model IDs for application inference profiles.

Currently, the `base_model_id` check and potential Bedrock
GetInferenceProfile API call are done inside the `_get_base_model`
method, which may cause a new API request to be executed on every
invocation of the chat model. We can instead relocate this mechanism to
the environment pre-validator method - this will ensure that we make the
extra API request only once, on instantiation of
ChatBedrock/ChatBedrockConverse.

---------

Co-authored-by: Piyush Jain <piyushjain@duck.com>