feat: Adding Bedrock AgentCore Memory Store for long term memory (#611)
## Summary

LangGraph memory stores are intended to be used hand in hand with
checkpointers to process and create long term memories from
conversations. This fits well into the Bedrock AgentCore Memory
framework where chat events are saved and processed async into long term
memories.


https://langchain-ai.github.io/langgraph/concepts/persistence/#memory-store

> With this all in place, we use the in_memory_store in LangGraph. The
in_memory_store works hand-in-hand with the checkpointer: the
checkpointer saves state to threads, as discussed above, and the
in_memory_store allows us to store arbitrary information for access
across threads. We compile the graph with both the checkpointer and the
in_memory_store as follows.

## Differences with LangGraph

So for this implementation, the `AgentCoreMemoryStore` was created as an
alternative to `InMemoryStore`. It overrides the `put`, `search`, `get`,
and `delete` methods to interact with Bedrock AgentCore Memory instead
of storing the messages in a list and directly performing the embedding.

There are some nuances with Bedrock AgentCore namespaces and the
difference between short term events and long term memory records. In
LangGraph, long term memories are stored directly and managed through
CRUD operations (with the help of an embedding model if the user
chooses). For this implementation, I played around with Bedrock
AgentCore events being saved with the `put` operation, then the `search`
operation will query long term memories or simply list long term
memories if no query is provided.

A sample notebook walks through usage of this implementation and how it
fits into a LangGraph agent.

---------

Co-authored-by: Jack Gordley <jgordle@amazon.com>
Co-authored-by: Piyush Jain <piyushjain@duck.com>