fix(aws): Handle thinking with tools, use CountTokens API for Haiku 4.5 (#698)
Couple of updates to ChatBedrock/ChatBedrockConverse for Haiku 4.5
support:
- Allowlisted Haiku 4.5 to use CountTokens API for
`get_num_tokens_from_messages()`.
- Disabled forced tool use + thinking with `bind_tools()` and
`with_structured_output()`, as Haiku (and Sonnet) 4.5 still do not
support this.
- Updated related unit/integration tests.
- **NOTE:** The
[`test_structured_output_tool_choice_not_supported`](https://github.com/langchain-ai/langchain-aws/blob/53298392adfbe2c02fcc14cf75383920b7935757/libs/aws/tests/integration_tests/chat_models/test_bedrock_converse.py#L436)
Converse integration test has not been updated to include Haiku 4.5,
which appears to behave very inconsistently with a structured output
schema compared to previous Claude thinking models. Sonnet 4.5 and older
models produce a tool call 100% of the time if the query is directly
applicable to the schema. However, on Haiku 4.5, the chance of this
happening was observed to be <5%.