fix(langgraph-checkpoint-aws): Add `max_results` and `limit` attributes to AgentCoreMemorySaver (#728)
Addresses #719.

This PR updates the AgentCore memory checkpointer integration to:
- Add an `max_results` attribute to `AgentCoreMemorySaver`, allowing
control of the `maxResults` value passed to [ListEvents
API](https://docs.aws.amazon.com/bedrock-agentcore/latest/APIReference/API_ListEvents.html).
- Effective default will remain at the previously hardcoded limit of 100
results.
- Add another `limit` attribute to control the maximum events parsed
when paginating on the ListEvents response.
  - Effective default will move from 100 -> `None` (unlimited).
- In `AgentCoreEventClient,get_events()`, exit from the response
processing loop early if the number of events found exceeds the `limit`
attribute value. If this occurs, a warning will be returned to the user.